server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
asyncmy>=0.2.4
python-dotenv>=0.19.0
pydantic>=1.8.2